Initial Push
A hybrid C/C++ program that simulates demand paging using six different
algorithms: FIFO, Least Recently Used, Most Recently Used, the
theoretical optimum algorithm, and then two algorithms that randomly
generate numbers: one that uses a uniformly distributed scheme of random
numbers to calculate frames that should be freed and one that uses the
pseudorandom rand() function supplied with C/C++ in order to calculate
pages to remove.

A hybrid of both languages was chosen due to the want for the speed and
preciseness that C offers with the added libraries and functions that
C++ provides.